Extract audio conversion into separate class (#3130)
Summary:
Pull Request resolved: https://github.com/pytorch/audio/pull/3130

Similar to https://github.com/pytorch/audio/pull/3120
Adopt the generator style slicing conversion to audio encoding
process.

Differential Revision: D43685380

fbshipit-source-id: 0e2f28cdb6bac55fb134a10229085ebf4e4c69a5
